v1/requirements.txt
flask
groq
pypdf
python-docx
